Evaluator/requirements.txt
asgiref==3.2.10
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==3.1
django-extensions==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.5
packaging==21.3
pandas==1.5.1
paramiko==2.12.0
patsy==0.5.3
Pillow==9.3.0
psycopg2==2.8.6
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
reportlab==3.6.12
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
statsmodels==0.13.5
threadpoolctl==3.1.0
unicode==2.9
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
